requirements.txt
python=3
keras
tensorflow-gpu
sklearn
nibabel
pandas